requirements.txt
python-telegram-bot==20.7
yfinance==0.2.36
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.4.0
